requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0
numpy>=1.24
psycopg2-binary>=2.9
python-dotenv>=1.0
openai>=1.30.0
